Remove RecordTransaction::terminate() method
This method leads to multiple Agent::send() calls under certain conditions. Furthermore, the Laravel documentation implies it is only called under certain circumstances. While useful in those cases, we cannot rely on it.

https://laravel.com/docs/7.x/middleware#terminable-middleware

"If you define a terminate method on your middleware and your web server is using FastCGI, the terminate method will automatically be called after the response is sent to the browser"